ETL/src/requirements.txt
apache-flink
redshift_connector
numpy
pandas
boto3
botocore
python-dotenv